refactor(ui): keep project helpers with task utilities
