+++ packages/ui/client/components/FileDetails.vue
@@ -14,8 +14,7 @@ import {
 import { tagsDefinitions } from '~/composables/client/state'
 import { hasFailedSnapshot } from '~/composables/explorer/collector'
 import { selectedTest, viewMode } from '~/composables/params'
-import { getProjectBadgeStyle } from '~/utils/project'
-import { getBadgeNameColor } from '~/utils/task'
+import { getBadgeNameColor, getProjectBadgeStyle } from '~/utils/task'
 import FileDetailsModuleGraph from './FileDetailsModuleGraph.vue'
 import IconButton from './IconButton.vue'
 import StatusIcon from './StatusIcon.vue'
