+++ packages/ui/client/utils/task.ts
@@ -1,4 +1,4 @@
-import type { RunnerTask, RunnerTestSuite } from 'vitest'
+import type { RunnerTask, RunnerTestSuite, SerializedRootConfig } from 'vitest'
 import { isDark } from '~/composables'
 
 export function isSuite(task: RunnerTask): task is RunnerTestSuite {
@@ -159,5 +159,26 @@ export function getBadgeTextColor(color: string) {
     case 'white':
     default:
       return 'black'
+  }
+}
+
+export function getProjectConfigByName(
+  // TODO: config shouldn't be partial in valid app lifetime. rework later.
+  config: Partial<SerializedRootConfig>,
+  projectName: string | undefined,
+) {
+  return config.projects?.find(project => project.name === projectName)
+}
+
+export function getProjectBadgeStyle(
+  config: Partial<SerializedRootConfig>,
+  projectName: string | undefined,
+) {
+  const backgroundColor = getProjectConfigByName(config, projectName)?.color
+    ?? getBadgeNameColor(projectName)
+
+  return {
+    backgroundColor,
+    color: getBadgeTextColor(backgroundColor),
   }
 }
